insert gray_r into zoedepth.infer
