+++ handyinfer/utils/__init__.py
@@ -1,3 +1,3 @@
-from .misc import TensorGrayR, img2tensor, load_file_from_url, scandir, tensor2img_fast
+from .misc import img2tensor, load_file_from_url, scandir, tensor2img_fast
 
-__all__ = ['load_file_from_url', 'img2tensor', 'scandir', 'tensor2img_fast', 'TensorGrayR']
+__all__ = ['load_file_from_url', 'img2tensor', 'scandir', 'tensor2img_fast']
